fix cheerpj site
