+++ packages/astro-theme/components/ProductVersionSwitcher.astro
@@ -24,14 +24,10 @@ const { product } = Astro.props;
 		| HTMLSelectElement
 		| undefined;
 	select?.addEventListener("change", async () => {
-		let href = location.pathname.replace(/\/cheerpj\d/, `/${select.value}`);
-
-		// If the page doesn't exist, redirect to the homepage
-		const req = await fetch(href, { method: "HEAD" });
-		if (req.status === 404) {
-			href = `/${select.value}`;
+		if (select.value === "cheerpj3") {
+			location.href = "https://cheerpj.com/docs";
+		} else if (select.value === "cheerpj2") {
+			location.href = "https://labs.leaningtech.com/docs/cheerpj2"
 		}
-
-		location.href = href;
 	});
 </script>
