+++ sites/cheerpj/src/content/docs/00-overview.mdx
@@ -1,5 +1,6 @@
 ---
-title: Overview
+title: CheerpJ
+shortTitle: Overview
 description: Java Virtual Machine for modern web browsers
 ---
