+++ sites/cheerpj/src/content/docs/24-community.mdx
@@ -17,10 +17,10 @@ Become a member of our community today!
 
 ## Projects built with CheerpJ
 
-The Leaning Technologies community has built some amazing projects with CheerpJ. Here are some of our favorites.
+The CheerpJ community has built some amazing projects. Here are some of our favorites.
 
-<ShowcaseList tag="CheerpJ" />
+<ShowcaseList />
 
 Are you using CheerpJ in a project? [Let us know][submit] and we'll feature it!
 
-[submit]: https://github.com/leaningtech/labs.leaningtech.com/issues/new?assignees=bates64&labels=documentation&projects=&template=project-submission.md&title=Project+submission
+[submit]: https://github.com/leaningtech/labs/issues/new?labels=documentation&projects=&template=project-submission.md&title=Project+submission
