+++ sites/labs/src/content/config.ts
@@ -42,18 +42,4 @@ const blog = defineCollection({
 
 const docs = defineDocsCollection();
 
-const showcase = defineCollection({
-	type: "data",
-	schema: ({ image }) =>
-		z.object({
-			title: z.string(),
-			description: z.string().optional(),
-			url: z.string(),
-			image: image().refine((img) => img.width / img.height == 1.5, {
-				message: "Image must have 3:2 aspect ratio",
-			}),
-			tags: productTags,
-		}),
-});
-
-export const collections = { docs, blog, blogauthors, showcase };
+export const collections = { docs, blog, blogauthors };
